requirements/common.txt
flask
flask-login
celery
redis
flask-wtf
flask-sqlalchemy
pip-tools
python-codicefiscale
email_validator
python-dotenv